requirements.txt
discord.py==2.5.0
openai==1.64.0
python-dotenv==1.0.1
requests==2.32.3